Move GameAssetManager to its own file. Verify and fix all functions in GameMemory.cpp including properly adding Create_W3D_Mem_Pool, Allocate_From_W3D_Mem_Pool and Free_From_W3D_Mem_Pool. Move array.h to w3d\lib. Create crcengine.cpp and realcrc.cpp and move things from crc.cpp as needed (and use in w3d where it should be used). Move some stuff from rtsutils.h to a new rtsutilsw3d.h in w3d\lib so that w3d can use that without dragging in other things. Correctly implement IMPLEMENT_NAMED_W3D_POOL macro.
