+++ src/game/common/crc.h
@@ -25,13 +25,8 @@ public:
     void Compute_CRC(void const *data, int bytes);
     uint32_t Get_CRC() { return m_crc; }
 
-    static uint32_t Memory(void const *data, size_t bytes, uint32_t crc);
-    static uint32_t String(const char *string, uint32_t crc);
-    static uint32_t Stringi(const char *string, uint32_t crc);
-
 private:
     void Add_CRC(uint8_t byte);
 
-    static const Array<uint32_t, 256> m_table;
     uint32_t m_crc;
 };
