+++ src/game/common/system/gamememory.h
@@ -27,16 +27,12 @@ extern bool g_thePreMainInitFlag;
 extern bool g_theMainInitFlag;
 #endif
 
-#ifdef GAME_DEBUG
-extern int g_theLinkChecker;
-#endif
-
 void Init_Memory_Manager();
 void Init_Memory_Manager_Pre_Main();
 void Shutdown_Memory_Manager();
-MemoryPool *Create_Named_Pool(const char *name, int size);
-void *Allocate_From_Pool(MemoryPool *pool, int size);
-void Free_From_Pool(MemoryPool *pool, void *memory);
+void *Create_W3D_Mem_Pool(char const *name, int size);
+void *Allocate_From_W3D_Mem_Pool(void *pool, int size);
+void Free_From_W3D_Mem_Pool(void *pool, void *data);
 
 #ifdef GAME_DLL
 void *New_New(size_t bytes);
