+++ src/w3d/lib/array.h
