+++ src/w3d/renderer/textureloader.cpp
@@ -19,7 +19,7 @@
 #include "ddsfile.h"
 #include "dx8wrapper.h"
 #include "missing.h"
-#include "rtsutils.h"
+#include "rtsutilsw3d.h"
 #include "shader.h"
 #include "synctextureloadtasklist.h"
 #include "targa.h"
